chore(cli): finalize runtime command decomposition
